+++ lib/compat/wordpress-6.0/class-wp-theme-json-gutenberg.php
@@ -1,0 +1,45 @@
+<?php
+/**
+ * WP_Theme_JSON_Gutenberg class
+ *
+ * @package gutenberg
+ */
+
+/**
+ * Class that encapsulates the processing of structures that adhere to the theme.json spec.
+ *
+ * This class is for internal core usage and is not supposed to be used by extenders (plugins and/or themes).
+ * This is a low-level API that may need to do breaking changes. Please,
+ * use get_global_settings, get_global_styles, and get_global_stylesheet instead.
+ *
+ * @access private
+ */
+class WP_Theme_JSON_Gutenberg extends WP_Theme_JSON_5_9 {
+
+	/**
+	 * The top-level keys a theme.json can have.
+	 *
+	 * @var string[]
+	 */
+	const VALID_TOP_LEVEL_KEYS = array(
+		'customTemplates',
+		'patterns',
+		'settings',
+		'styles',
+		'templateParts',
+		'version',
+	);
+
+	/**
+	 * Returns the current theme's wanted patterns(slugs) to be
+	 * registered from Pattern Directory.
+	 *
+	 * @return array
+	 */
+	public function get_patterns() {
+		if ( isset( $this->theme_json['patterns'] ) && is_array( $this->theme_json['patterns'] ) ) {
+			return $this->theme_json['patterns'];
+		}
+		return array();
+	}
+}