+++ packages/block-library/src/query-title/index.js
@@ -6,10 +6,16 @@ import edit from './edit';
 import variations from './variations';
 import deprecated from './deprecated';
 
+/**
+ * WordPress dependencies
+ */
+import { queryTitle as icon } from '@wordpress/icons';
+
 const { name } = metadata;
 export { metadata, name };
 
 export const settings = {
+	icon,
 	edit,
 	variations,
 	deprecated,