+++ packages/e2e-tests/specs/site-editor/template-part.test.js
@@ -3,18 +3,14 @@
  */
 import {
 	insertBlock,
-	trashAllPosts,
+	deleteAllTemplates,
 	activateTheme,
 	getAllBlocks,
 	selectBlockByClientId,
 	clickBlockToolbarButton,
 	canvas,
+	visitSiteEditor,
 } from '@wordpress/e2e-test-utils';
-
-/**
- * Internal dependencies
- */
-import { siteEditor } from './utils';
 
 const templatePartNameInput =
 	'.edit-site-create-template-part-modal .components-text-control__input';
@@ -22,24 +18,23 @@ const templatePartNameInput =
 describe( 'Template Part', () => {
 	beforeAll( async () => {
 		await activateTheme( 'emptytheme' );
-		await trashAllPosts( 'wp_template' );
-		await trashAllPosts( 'wp_template_part' );
+		await deleteAllTemplates( 'wp_template' );
+		await deleteAllTemplates( 'wp_template_part' );
 	} );
 	afterAll( async () => {
-		await trashAllPosts( 'wp_template' );
-		await trashAllPosts( 'wp_template_part' );
+		await deleteAllTemplates( 'wp_template' );
+		await deleteAllTemplates( 'wp_template_part' );
 		await activateTheme( 'twentytwentyone' );
 	} );
 
 	describe( 'Template part block', () => {
 		beforeEach( async () => {
-			await siteEditor.visit();
-			await siteEditor.disableWelcomeGuide();
+			await visitSiteEditor();
 		} );
 
 		async function navigateToHeader() {
 			// Switch to editing the header template part.
-			await siteEditor.visit( {
+			await visitSiteEditor( {
 				postId: 'emptytheme//header',
 				postType: 'wp_template_part',
 			} );
@@ -60,7 +55,7 @@ describe( 'Template Part', () => {
 			);
 
 			// Switch back to the Index template.
-			await siteEditor.visit( {
+			await visitSiteEditor( {
 				postId: 'emptytheme//index',
 				postType: 'wp_template',
 			} );
@@ -314,7 +309,7 @@ describe( 'Template Part', () => {
 				await page.click( entitiesSaveSelector );
 
 				// Reload the page so as the new template part is available in the existing template parts.
-				await siteEditor.visit();
+				await visitSiteEditor();
 				siteEditorCanvas = canvas();
 				await awaitHeaderLoad();
 				// Try to insert the template part we created.