+++ packages/data/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 6.2.1 (2022-02-10)
+
+### Bug Fix
+
+-   Removed unused `memize` dependency ([#38388](https://github.com/WordPress/gutenberg/pull/38388)).
+
 ## 6.2.0 (2022-01-27)
 
 ### Bug Fix